Pascal Triangle Using Recursion
We’ve used a nested list as the data structure, where each row in Pascal’s triangle is a list in itself, like this: [[row 1], [row 2],…,[row n]].
The function call pas_tri(numRows) triggers a series of recursive calls with numRows - 1, numRows - 2 all the way up to 1 as the arguments. These calls are pushed onto a stack.
When numRows == 1, we’ve reached the base case, and the function returns [[1]].
Now the returned list is used by the subsequent functions in the call stack—to compute the next row.
If cur_row is the current row, cur_row[i] = prev_row[i] + prev_row[i+1]—the sum of 2 elements directly above the current index.